fix bench shell

+++ isuumo/bench/restart-and-bench.sh
@@ -9,7 +9,12 @@ sleep 5
 echo | sudo tee /var/log/nginx/access.log
 
 ./bench -target-url http://127.0.0.1:8080
+
 chmod 777 /var/log/nginx/access.log
-rm alp_kekka.txt
+if [ -e alp_kekka.txt ]; then
+ rm alp_kekka.txt
+fi
+
+sleep 3
 /var/log/nginx/access.log | alp ltsv --sort avg -r >> alp_kekka.txt
